+++ src/Storages/ObjectStorage/DataLakes/Iceberg/IcebergMetadata.cpp
@@ -87,18 +87,14 @@ IcebergMetadata::IcebergMetadata(
     ObjectStoragePtr object_storage_,
     StorageObjectStorageConfigurationWeakPtr configuration_,
     const ContextPtr & context_,
-    Int32 metadata_version_,
     Int32 format_version_,
-    const Poco::JSON::Object::Ptr & metadata_object_,
     IcebergMetadataFilesCachePtr cache_ptr)
     : object_storage(std::move(object_storage_))
     , configuration(std::move(configuration_))
     , schema_processor(IcebergSchemaProcessor())
+    , iceberg_metadata_cache(cache_ptr)
     , log(getLogger("IcebergMetadata"))
-    , manifest_cache(cache_ptr)
-    , last_metadata_version(metadata_version_)
     , format_version(format_version_)
-    , relevant_snapshot_schema_id(-1)
     , table_location(metadata_object_->getValue<String>(f_location))
 {
     updateState(context_, metadata_object_, true);
@@ -227,29 +223,17 @@ bool IcebergMetadata::update(const ContextPtr & local_context)
     std::lock_guard lock(mutex);
 
     const auto [metadata_version, metadata_file_path, compression_method]
-        = getLatestOrExplicitMetadataFileAndVersion(object_storage, configuration_ptr, manifest_cache, local_context, log.get());
+        = getLatestOrExplicitMetadataFileAndVersion(object_storage, configuration_ptr, iceberg_metadata_cache, local_context, log.get());
 
-    bool metadata_file_changed = false;
-    if (last_metadata_version != metadata_version)
-    {
-        last_metadata_version = metadata_version;
-        metadata_file_changed = true;
-    }
-
-    auto metadata_object = getMetadataJSONObject(metadata_file_path, object_storage, configuration_ptr, manifest_cache, local_context, log, compression_method);
+    auto metadata_object = getMetadataJSONObject(
+        metadata_file_path, object_storage, configuration_ptr, iceberg_metadata_cache, local_context, log, compression_method);
     chassert(format_version == metadata_object->getValue<int>(f_format_version));
 
     auto previous_snapshot_id = relevant_snapshot_id;
     auto previous_snapshot_schema_id = relevant_snapshot_schema_id;
 
-    updateState(local_context, metadata_object, metadata_file_changed);
+    updateState(local_context, metadata_object);
 
-    if (previous_snapshot_id != relevant_snapshot_id)
-    {
-        schema_id_by_data_file.clear();
-        schema_id_by_data_files_initialized = false;
-        return true;
-    }
     return previous_snapshot_schema_id != relevant_snapshot_schema_id;
 }
 
@@ -324,10 +308,15 @@ void IcebergMetadata::updateSnapshot(ContextPtr local_context, Poco::JSON::Objec
             }
 #endif
 
-            relevant_snapshot = IcebergSnapshot{
-                getManifestList(local_context, getProperFilePathFromMetadataInfo(
-                    snapshot->getValue<String>(f_manifest_list), configuration_ptr->getPathForRead().path, table_location)),
-                relevant_snapshot_id, total_rows, total_bytes, total_position_deletes};
+            relevant_snapshot = IcebergDataSnapshot{
+                getManifestList(
+                    local_context,
+                    getProperFilePathFromMetadataInfo(
+                        snapshot->getValue<String>(f_manifest_list), configuration_ptr->getPathForRead().path, table_location)),
+                relevant_snapshot_id,
+                total_rows,
+                total_bytes,
+                total_position_deletes};
 
             if (!snapshot->has(f_schema_id))
                 throw Exception(
@@ -402,49 +391,6 @@ void IcebergMetadata::updateState(const ContextPtr & local_context, Poco::JSON::
     }
 }
 
-std::shared_ptr<NamesAndTypesList> IcebergMetadata::getInitialSchemaByPath(ContextPtr local_context, const String & data_path) const
-{
-    if (!schema_id_by_data_files_initialized)
-    {
-        std::lock_guard lock(mutex);
-        if (!schema_id_by_data_files_initialized)
-            initializeSchemasFromManifestList(local_context, relevant_snapshot->manifest_list_entries);
-    }
-
-    SharedLockGuard lock(mutex);
-    auto version_if_outdated = getSchemaVersionByFileIfOutdated(data_path);
-    return version_if_outdated.has_value() ? schema_processor.getClickhouseTableSchemaById(version_if_outdated.value()) : nullptr;
-}
-
-std::shared_ptr<const ActionsDAG> IcebergMetadata::getSchemaTransformer(ContextPtr local_context, const String & data_path) const
-{
-    if (!schema_id_by_data_files_initialized)
-    {
-        std::lock_guard lock(mutex);
-        if (!schema_id_by_data_files_initialized)
-            initializeSchemasFromManifestList(local_context, relevant_snapshot->manifest_list_entries);
-    }
-
-    SharedLockGuard lock(mutex);
-    auto version_if_outdated = getSchemaVersionByFileIfOutdated(data_path);
-    return version_if_outdated.has_value()
-        ? schema_processor.getSchemaTransformationDagByIds(version_if_outdated.value(), relevant_snapshot_schema_id)
-        : nullptr;
-}
-
-std::optional<Int32> IcebergMetadata::getSchemaVersionByFileIfOutdated(String data_path) const
-{
-    auto schema_id_it = schema_id_by_data_file.find(data_path);
-    if (schema_id_it == schema_id_by_data_file.end())
-        throw Exception(ErrorCodes::BAD_ARGUMENTS, "Cannot find manifest file for data file: {}", data_path);
-
-    auto schema_id = schema_id_it->second;
-    if (schema_id == relevant_snapshot_schema_id)
-        return std::nullopt;
-
-    return std::optional{schema_id};
-}
-
 void IcebergMetadata::createInitial(
     const ObjectStoragePtr & object_storage,
     const StorageObjectStorageConfigurationWeakPtr & configuration,
@@ -516,24 +462,6 @@ DataLakeMetadataPtr IcebergMetadata::create(
     return std::make_unique<IcebergMetadata>(object_storage, configuration_ptr, local_context, metadata_version, format_version, object, cache_ptr);
 }
 
-void IcebergMetadata::initializeSchemasFromManifestList(ContextPtr local_context, ManifestFileCacheKeys manifest_list_ptr) const
-{
-    schema_id_by_data_file.clear();
-
-    for (const auto & manifest_list_entry : manifest_list_ptr)
-    {
-        auto manifest_file_ptr = getManifestFile(
-            local_context,
-            manifest_list_entry.manifest_file_path,
-            manifest_list_entry.added_sequence_number,
-            manifest_list_entry.added_snapshot_id);
-        for (const auto & manifest_file_entry : manifest_file_ptr->getFiles(Iceberg::FileContentType::DATA))
-        {
-            schema_id_by_data_file.emplace(manifest_file_entry.file_path, manifest_file_entry.schema_id);
-        }
-    }
-}
-
 ManifestFileCacheKeys IcebergMetadata::getManifestList(ContextPtr local_context, const String & filename) const
 {
     auto configuration_ptr = configuration.lock();
@@ -546,7 +474,7 @@ ManifestFileCacheKeys IcebergMetadata::getManifestList(ContextPtr local_context,
 
         auto read_settings = local_context->getReadSettings();
         /// Do not utilize filesystem cache if more precise cache enabled
-        if (manifest_cache)
+        if (iceberg_metadata_cache)
             read_settings.enable_filesystem_cache = false;
 
         auto manifest_list_buf = StorageObjectStorageSource::createReadBuffer(object_info, object_storage, local_context, log, read_settings);
@@ -580,8 +508,9 @@ ManifestFileCacheKeys IcebergMetadata::getManifestList(ContextPtr local_context,
     };
 
     ManifestFileCacheKeys manifest_file_cache_keys;
-    if (manifest_cache)
-        manifest_file_cache_keys = manifest_cache->getOrSetManifestFileCacheKeys(IcebergMetadataFilesCache::getKey(configuration_ptr, filename), create_fn);
+    if (iceberg_metadata_cache)
+        manifest_file_cache_keys = iceberg_metadata_cache->getOrSetManifestFileCacheKeys(
+            IcebergMetadataFilesCache::getKey(configuration_ptr, filename), create_fn);
     else
         manifest_file_cache_keys = create_fn();
     return manifest_file_cache_keys;
@@ -591,15 +520,12 @@ IcebergMetadata::IcebergHistory IcebergMetadata::getHistory(ContextPtr local_con
 {
     auto configuration_ptr = configuration.lock();
 
-    const auto [metadata_version, metadata_file_path, compression_method] = getLatestOrExplicitMetadataFileAndVersion(object_storage, configuration_ptr, manifest_cache, local_context, log.get());
 
-    chassert([&]()
-    {
-        SharedLockGuard lock(mutex);
-        return metadata_version == last_metadata_version;
-    }());
+    const auto [metadata_version, metadata_file_path, compression_method]
+        = getLatestOrExplicitMetadataFileAndVersion(object_storage, configuration_ptr, iceberg_metadata_cache, local_context, log.get());
 
-    auto metadata_object = getMetadataJSONObject(metadata_file_path, object_storage, configuration_ptr, manifest_cache, local_context, log, compression_method);
+    auto metadata_object = getMetadataJSONObject(
+        metadata_file_path, object_storage, configuration_ptr, iceberg_metadata_cache, local_context, log, compression_method);
     chassert([&]()
     {
         SharedLockGuard lock(mutex);
@@ -677,134 +603,7 @@ IcebergMetadata::IcebergHistory IcebergMetadata::getHistory(ContextPtr local_con
     return iceberg_history;
 }
 
-ManifestFilePtr IcebergMetadata::getManifestFile(
-    ContextPtr local_context, const String & filename, Int64 inherited_sequence_number, Int64 inherited_snapshot_id) const
-{
-    auto configuration_ptr = configuration.lock();
-
-    auto create_fn = [&]()
-    {
-        ObjectInfo manifest_object_info(filename);
-
-        auto read_settings = local_context->getReadSettings();
-        /// Do not utilize filesystem cache if more precise cache enabled
-        if (manifest_cache)
-            read_settings.enable_filesystem_cache = false;
-
-        auto buffer = StorageObjectStorageSource::createReadBuffer(manifest_object_info, object_storage, local_context, log, read_settings);
-        AvroForIcebergDeserializer manifest_file_deserializer(std::move(buffer), filename, getFormatSettings(local_context));
-
-        return std::make_shared<ManifestFileContent>(
-            manifest_file_deserializer,
-            filename,
-            format_version,
-            configuration_ptr->getPathForRead().path,
-            schema_processor,
-            inherited_sequence_number,
-            inherited_snapshot_id,
-            table_location,
-            local_context);
-    };
-
-    if (manifest_cache)
-    {
-        auto manifest_file
-            = manifest_cache->getOrSetManifestFile(IcebergMetadataFilesCache::getKey(configuration_ptr, filename), create_fn);
-        return manifest_file;
-    }
-    return create_fn();
-}
-
-
 // We need to pass transform function here not to store ManifestFileEntry for data files explicitly in RAM
-template <typename T>
-std::vector<T> IcebergMetadata::getFilesImpl(
-    const ActionsDAG * filter_dag,
-    FileContentType file_content_type,
-    ContextPtr local_context,
-    std::function<T(const ManifestFileEntry &)> transform_function) const
-{
-    if (!local_context && filter_dag)
-    {
-        throw DB::Exception(
-            DB::ErrorCodes::LOGICAL_ERROR,
-            "Context is required with non-empty filter_dag to implement partition pruning for Iceberg table");
-    }
-
-    bool use_partition_pruning = filter_dag && local_context->getSettingsRef()[Setting::use_iceberg_partition_pruning].value;
-
-    std::vector<T> files;
-    {
-        SharedLockGuard lock(mutex);
-
-        if (!relevant_snapshot)
-            return {};
-
-
-        for (const auto & manifest_list_entry : relevant_snapshot->manifest_list_entries)
-        {
-            Int64 previous_entry_schema = -1;
-            std::optional<ManifestFilesPruner> pruner;
-            auto manifest_file_ptr = getManifestFile(
-                local_context,
-                manifest_list_entry.manifest_file_path,
-                manifest_list_entry.added_sequence_number,
-                manifest_list_entry.added_snapshot_id);
-            const auto & data_files_in_manifest = manifest_file_ptr->getFiles(file_content_type);
-            for (const auto & manifest_file_entry : data_files_in_manifest)
-            {
-                // Trying to reuse already initialized pruner
-                if ((manifest_file_entry.schema_id != previous_entry_schema) && (use_partition_pruning))
-                {
-                    previous_entry_schema = manifest_file_entry.schema_id;
-                    if (previous_entry_schema > manifest_file_entry.schema_id)
-                    {
-                        LOG_WARNING(log, "Manifest entries in file {} are not sorted by schema id", manifest_list_entry.manifest_file_path);
-                    }
-                    pruner.emplace(
-                        schema_processor,
-                        relevant_snapshot_schema_id,
-                        manifest_file_entry.schema_id,
-                        filter_dag ? filter_dag : nullptr,
-                        *manifest_file_ptr,
-                        local_context);
-                }
-
-                if (manifest_file_entry.status != ManifestEntryStatus::DELETED)
-                {
-                    if (!use_partition_pruning || !pruner->canBePruned(manifest_file_entry))
-                    {
-                        files.push_back(transform_function(manifest_file_entry));
-                    }
-                }
-            }
-        }
-    }
-    std::sort(files.begin(), files.end());
-    return files;
-}
-
-std::vector<ParsedDataFileInfo> IcebergMetadata::getDataFiles(
-    const ActionsDAG * filter_dag, ContextPtr local_context, const std::vector<ManifestFileEntry> & position_delete_files) const
-{
-    return getFilesImpl<ParsedDataFileInfo>(
-        filter_dag,
-        FileContentType::DATA,
-        local_context,
-        [this, &position_delete_files](const ManifestFileEntry & entry)
-        { return ParsedDataFileInfo{this->configuration.lock(), entry, position_delete_files}; });
-}
-
-std::vector<Iceberg::ManifestFileEntry>
-IcebergMetadata::getPositionDeleteFiles(const ActionsDAG * filter_dag, ContextPtr local_context) const
-{
-    return getFilesImpl<ManifestFileEntry>(
-        filter_dag,
-        FileContentType::POSITION_DELETE,
-        local_context,
-        // In the current design we can't avoid storing ManifestFileEntry in RAM explicitly for position deletes
-        [](const ManifestFileEntry & entry) { return entry; });
-}
 
 std::optional<size_t> IcebergMetadata::totalRows(ContextPtr local_context) const
 {
@@ -898,12 +697,6 @@ NamesAndTypesList IcebergMetadata::getTableSchema() const
     return *schema_processor.getClickhouseTableSchemaById(relevant_snapshot_schema_id);
 }
 
-std::tuple<Int64, Int32> IcebergMetadata::getVersion() const
-{
-    SharedLockGuard lock(mutex);
-    return std::make_tuple(relevant_snapshot_id, relevant_snapshot_schema_id);
-}
-
 bool IcebergMetadata::hasPositionDeleteTransformer(const ObjectInfoPtr & object_info) const
 {
     auto iceberg_object_info = std::dynamic_pointer_cast<IcebergDataObjectInfo>(object_info);
@@ -934,49 +727,6 @@ std::shared_ptr<ISimpleTransform> IcebergMetadata::getPositionDeleteTransformer(
 
     return std::make_shared<IcebergBitmapPositionDeleteTransform>(
         header, iceberg_object_info, object_storage, format_settings, context_, delete_object_format, delete_object_compression_method);
-}
-
-ParsedDataFileInfo::ParsedDataFileInfo(
-    StorageObjectStorageConfigurationPtr configuration_,
-    Iceberg::ManifestFileEntry data_object_,
-    const std::vector<Iceberg::ManifestFileEntry> & position_deletes_objects_)
-    : data_object_file_path_key(data_object_.file_path_key)
-    , data_object_file_path(data_object_.file_path)
-{
-    ///Object in position_deletes_objects_ are sorted by common_partition_specification, partition_key_value and added_sequence_number.
-    /// It is done to have an invariant that position deletes objects which corresponds
-    /// to the data object form a subsegment in a position_deletes_objects_ vector.
-    /// We need to take all position deletes objects which has the same partition schema and value and has added_sequence_number
-    /// greater than or equal to the data object added_sequence_number (https://iceberg.apache.org/spec/#scan-planning)
-    /// ManifestFileEntry has comparator by default which helps to do that.
-    auto beg_it = std::lower_bound(position_deletes_objects_.begin(), position_deletes_objects_.end(), data_object_);
-    auto end_it = std::upper_bound(
-        position_deletes_objects_.begin(),
-        position_deletes_objects_.end(),
-        data_object_,
-        [](const Iceberg::ManifestFileEntry & lhs, const Iceberg::ManifestFileEntry & rhs)
-        {
-            return std::tie(lhs.common_partition_specification, lhs.partition_key_value)
-                < std::tie(rhs.common_partition_specification, rhs.partition_key_value);
-        });
-    if (beg_it - position_deletes_objects_.begin() > end_it - position_deletes_objects_.begin())
-    {
-        throw Exception(
-            ErrorCodes::LOGICAL_ERROR,
-            "Position deletes objects are not sorted by common_partition_specification and partition_key_value, "
-            "beginning: {}, end: {}, position_deletes_objects size: {}",
-            beg_it - position_deletes_objects_.begin(),
-            end_it - position_deletes_objects_.begin(),
-            position_deletes_objects_.size());
-    }
-    position_deletes_objects = std::span<const Iceberg::ManifestFileEntry>{beg_it, end_it};
-    if (!position_deletes_objects.empty() && configuration_->format != "Parquet")
-    {
-        throw Exception(
-            ErrorCodes::UNSUPPORTED_METHOD,
-            "Position deletes are only supported for data files of Parquet format in Iceberg, but got {}",
-            configuration_->format);
-    }
 }
