+++ src/Storages/ObjectStorage/DataLakes/Iceberg/Snapshot.h
@@ -10,13 +10,14 @@
 namespace Iceberg
 {
 
-struct IcebergSnapshot
+struct IcebergDataSnapshot
 {
     DB::ManifestFileCacheKeys manifest_list_entries;
     Int64 snapshot_id;
     std::optional<size_t> total_rows;
     std::optional<size_t> total_bytes;
     std::optional<size_t> total_position_delete_rows;
+    Int32 schema_id;
 
     std::optional<size_t> getTotalRows() const
     {
@@ -25,6 +26,15 @@ struct IcebergSnapshot
         return std::nullopt;
     }
 };
+
+struct IcebergTableStateSnapshot
+{
+    Int32 metadata_version;
+    std::optional<Int32> snapshot_id;
+    std::optional<Int32> schema_id; // Set if snapshot is null or schema was changed after corresponding data commit
+};
+
+using IcebergTableStateSnapshotPtr = std::shared_ptr<IcebergTableStateSnapshot>;
 
 struct IcebergHistoryRecord
 {
